Preferences: Lift out PluginSliderWithInput from the MouseKeys prefs
